\name{nhppSpikeConstSignal}
\alias{nhppSpikeConstSignal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Spike NHPP rate with constant signal strength
}
\description{
Randomly spike the smoothed control rate of an NHPP according to the parameters, with constant signal strength.
}
\usage{
nhppSpikeConstSignal(smoothRates, nSpike = 25, cptLenR = 4, cptLenMean = 10, gainRate = 3, lossRate = 0.3, pGain = 0.6)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{smoothRates}{
The smoothed rate estimate of the control process
}
  \item{nSpike}{
Number of signal spikes
}
  \item{cptLenR}{
Parameter for signal width (parameter R of negative binomial)
}
  \item{cptLenMean}{
Parameter for signal width (mean width)
}
  \item{gainRate}{
Relative CN Gain
}
  \item{lossRate}{
Relative CN Loss
}
  \item{pGain}{
Proportion of signal regions being CN Gain
}
}
\details{
The signal width is randomly generated by negative binomial distribution with the two parameters given. This is a more restricted version of \code{\link{nhppSpike}}
}
\value{
  \item{spikeMat }{A matrix containing the actual signal spike information}
  \item{caseRates }{The rate of the case PP to be simulated after signal spike}
}
\author{
Jeremy J. Shen
}

\seealso{
\code{\link{nhppSimulate}}
}