\name{readSeqJShen}
\alias{readSeqJShen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Read my private data format
}
\description{
Read a private data format of J. Shen. Not to be used.
}
\usage{
readSeqJShen(filename)
}

\arguments{
  \item{filename}{
File name of the dataset
}
}
\details{
Read a private data format of J. Shen. Not to be used.
}
\value{
  \item{seqF }{Read position for each read}
  \item{seqChr }{Chromosome of each mapped read}
}
\author{
Jeremy J. Shen
}

\seealso{
\code{\link{readSeq}}, \code{\link{readSeqChiang}}, \code{\link{readSeqELANDPaired}}
}