\name{readSeqMLove}
\alias{readSeqMLove}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Read Mike Love's private data format
}
\description{
Read a private data format generated by a script of Mike Love. Not to be used.
}
\usage{
readSeqMLove(filename)
}

\arguments{
  \item{filename}{
File name of the dataset
}
}
\details{
Read a private data format generated by a script of Mike Love. Not to be used.
}
\value{
  \item{seqF }{Read position for each read}
  \item{seqChr }{Chromosome of each mapped read}
}
\author{
Jeremy J. Shen
}

\seealso{
\code{\link{readSeq}}, \code{\link{readSeqChiang}}, \code{\link{readSeqELANDPaired}}
}